requirements_clean.txt
fastapi==0.116.1
uvicorn==0.40.0
python-multipart==0.0.20
python-dotenv==1.1.1
fastapi-mail==1.4.1

# Whisper dependencies
tqdm==4.66.4
tiktoken==0.7.0
regex==2024.11.6

# Pydantic
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydantic-extra-types==2.10.5

# Email
email-validator==2.3.0
cryptography==46.0.3

# HTTP
requests==2.32.3
urllib3==2.2.3
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10

# FastAPI internals
starlette==0.37.2
click==8.1.8
h11==0.16.0

# Gemini
google-generativeai==0.7.2
google-ai-generativelanguage==0.6.6
google-api-core==2.29.0
googleapis-common-protos==1.72.0
grpcio==1.62.3
grpcio-status==1.62.3
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.1
